Introduced "inert" write_lock
This is for operations which do not change the observable state of the Population. In the new model we can not override reserve() in OrdPopulation which would lead to needless re-sortings. With the new mechanism we can mark BasePopulation::reserve() as a non-modifying operation. This will also allow other functions not implemented now, like shrink_to_fit().